Architecture -  TourDetailsViewModel implemented successfully. Database connection and insert, delete, getAll functions successfully added. Camp and Tour models map to string and reverse converting successfully created for database connection

+++ app/src/main/java/com/natiqhaciyef/voyagers/view/viewmodel/tour/TourViewModel.kt
@@ -39,7 +39,7 @@ class TourViewModel @Inject constructor(
                     list.clear()
 
                     for (doc in docs) {
-                        val id = doc["id"].toString()
+                        val id = doc["id"].toString().toInt()
                         var name = doc["name"] as String
                         var image = doc["image"] as MutableList<String>
                         var info = doc["info"] as String
